Add Player's Test Assets
Add the player's test assets because test need to have a fixture number
of test data to test against so that what is expecting is correct even
if `rushing.json` changes. Furthermore, this can allow multiple versions
of the same test file. So that other types of tests can be performed
without having to change data in `rushing.json`.