Add React/TypeScript From Create React-App
Add React/TypeScript from the Create React-App so that the basic setup
is ready to go without having to do extra work to add all of the
packages that are now required to get a frontend working. Some of those
packages include ESLint, prettier, jest, webpack, and babel, to name a
few.

Also, change the port to 8080 so to not confliect with Rails port of
3000.