Add Jest Env JSDOM Sixteen to Fix MutationObserver
Add `jest-environment-jsdom-sixteen` as Testing Library has to remove
MutationObserver shim since the significant browsers and recent version
of JSDOM support MutationObserver, so there is no need for Testing
Library have their version of MutationObserver. However, `react-scripts`
have not updated to account for the loss of MutationObserver. As a
result, we add in manually until `react-scripts` has updated.

---
Ref:

https://github.com/testing-library/dom-testing-library/issues/477
https://github.com/testing-library/dom-testing-library/releases/tag/v7.0.0