Add Sorting to Players Model
Add the sorting to the Players model so that the frontend can choose how
it should display the data to the user. Also, the reason that the
sorting is happening at the array level instead of the database is that
Rails still does not check for SQL injection when running queries.

Since the frontend control what sorted, a bad actor could send a SQL
statement into the order statement and attack in that vector. Having the
order by at the array level lowerest the change that a SQL injection
attack could happen.