Add Fetch Data Hook to Get Data From Server
Add the Fetch Data hook to get data from the server and separate the
fetching from the presentation layer so that the presentation layer is
not tied down with a particular pattern and can move around. The same
goes for the Fetch Data hook so that it is not tied down to a specific
project and can be used to throw out this application without much
setup.