Add Global State Management
Add a global state management because with the higher-order component
 `WithFetchPlayersData`, it is clear that too much business logic is in
that component. As a result, that component has become bloated. To stop
this and bring some structure back to the frontend, moving the context
part of another file.

Also, creating proper global state management to handle the data coming
to form the sever, as well as event from both the search box and heading
sorting.