Update WithFetchPlayersData to Not be Context
Update the WithFetchPlayersData component not to be a context provider
as the global state take the data that came from the server and sends
them the components that subscribe to that data. Since the server URL is
part of the global state any time, the URL is updated. A fetch happens
and gets new data from the server.