Remove `useFetchData` Hook
Remove `useFetchData` hook as the Fetch Data component does the same
function but allows for JSX to be rendered as well. Furthermore, since
the application uses a global state and that there is one part of the
app that connects to the server, having a fetch hook would be more
complicated then it needs to be.