Fix No Player & Download Displays on First Loading
Fix an error where the text No Player and Download displays on first
loading only the loading text should appear when we are getting data
from the server. As a result of testing, the is a bug found in the test
where the data use not clear thus, unforeseen errors happen since that
data never got used during its test, which seems like a bug.