Add Pagination to Players Contoller & Model
Add pagination to the Player controller and model, not all of the
players to the frontend at the same time, which would slow the frontend
down as the number of players grows.

Since Kaminari gem is in ActiveRecord record, switch the sorting from
array to ActiveRecord using `order`. As a result, the sorting is
happening in the database.