Fix No Player Message on Error Message
Fix an issue where the No Player message would display when an error
happens. To not confuses the user and not have the screen be messy with
too many words. If there is an error, the error message should appear
only.

This fix sorting specs in App was always erroring out but was not cough,
which means they broken in this commit.