Fix App Test Sorting Not Working
Fix App test sorting not working since sorting test does not need to
know the data since the server takes care of the sorting and the
frontend display what is coming back. As a result, the mock value
happens only once, and no simulated data use again if more data calls
are needed, which results in the sorting not working.