Refactor Update URL to Reduce Code & Redirection
Refactor the updating the URL, thus the call to the server to reduce the
code and redirection of the code. Since the updating URL requires a call
function before dispatching. As a result, it creates overhead that does
not need to be there both for developers, having to go to the file see
the call and performance as every call has an extra step.