Add Code Coverage Ignore to Create Context
Add code coverage ignore to create context as there is no way for
developers nor the user to access the dispatch directly. As a result, no
test, require to test that path so that this line can safely ignore.

With this update, code coverage on the frontend is now at 100%.